+++ mellon-client/shell.nix
@@ -4,7 +4,6 @@ let
   haskellPackages = pkgs.haskellPackages.override {
     overrides = self: super: {
       servant-client = if pkgs.stdenv.isDarwin then pkgs.haskell.lib.dontCheck super.servant-client else super.servant-client;
-      mellon = pkgs.haskell.lib.dontCheck (pkgs.haskell.lib.appendConfigureFlag (self.callPackage ../mellon {}) "--ghc-options=-Werror");
       mellon-core = pkgs.haskell.lib.dontCheck (pkgs.haskell.lib.appendConfigureFlag (self.callPackage ../mellon-core {}) "--ghc-options=-Werror");
       mellon-server = pkgs.haskell.lib.dontCheck (pkgs.haskell.lib.appendConfigureFlag (self.callPackage ../mellon-server {}) "--ghc-options=-Werror");
       mellon-gpio = pkgs.haskell.lib.dontCheck (pkgs.haskell.lib.appendConfigureFlag (self.callPackage ../mellon-gpio {}) "--ghc-options=-Werror");