+++ mellon-server/changelog.md
@@ -1,3 +1,4 @@
 ## 0.7.0.0 (2016-06-02)
 
-Fix Servant bitrot.
+- Port to new `mellon-core` package.
+- Fix Servant bitrot.